Moving utility functions out of OkHttpDelivery class
