+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/DeliveryHelper.kt
@@ -24,7 +24,6 @@ object DeliveryHelper {
         return truncateEvent(payload, apiKey, maxStringValueLength, logger)
     }
 
-    @Suppress("MagicNumber")
     fun getDeliveryStatus(responseCode: Int): DeliveryStatus {
         return when {
             responseCode in HTTP_OK..299 -> DeliveryStatus.DELIVERED
@@ -33,7 +32,6 @@ object DeliveryHelper {
         }
     }
 
-    @Suppress("MagicNumber")
     fun isUnrecoverableStatusCode(responseCode: Int) =
         responseCode in HTTP_BAD_REQUEST..499 && // 400-499 are considered unrecoverable
             responseCode != HTTP_CLIENT_TIMEOUT && // except for 408
