+++ config/detekt/detekt.yml
@@ -30,6 +30,8 @@ style:
   ReturnCount:
     active: true
     max: 4
+  MagicNumber:
+    active: false
 
 # allow for utility function naming in OkHttpDelivery
 naming:
